Add files via upload
Updated version of the Robot Ethology GUI, which is currently required to be separate from the Novel Behavior code due to a difference in light sensor behavior.  The two robots -1894-wombat and 1897-wombat - currently used for the Ethology portion of the lab sequence have light sensors that behave opposite those used in Novel Behavior.  For the GUI version of the code, low light corresponds to low values, higher intensities corresponding to higher values.